backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.4
django-braces==1.14.0
django-compressor==2.4.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-dotenv==1.4.2
django-extensions==3.1.3
django-extra-views==0.14.0
django-jet-reboot==1.1.1
django-robots==4.0
django-sass-processor==1.0.1
django-silk==4.1.0
django-sql-explorer==2.3
django-taggit==1.5.1
django-taggit-serializer==0.1.7
django-unicorn==0.31.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
google-api-python-client==1.4.1
gprof2dot==2021.2.21
httplib2==0.19.1
idna==3.2
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
libsass==0.21.0
MarkupSafe==2.0.1
mock==4.0.3
nose==1.3.7
nplusone==1.0.0
oauth2client==4.1.3
oauthlib==3.1.1
orjson==3.6.0
packaging==21.0
Pillow==8.3.1
psutil==5.8.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pygraphviz==1.7
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
